revert to 16bit int for index for settimestamp

+++ r_code/utils.cpp
@@ -92,7 +92,7 @@ void Utils::SetTimestamp(Atom *iptr, uint64_t t) {
     iptr[2].atom = t & 0x00000000FFFFFFFF;
 }
 
-void Utils::SetTimestamp(Code *object, uint32_t index, uint64_t t)
+void Utils::SetTimestamp(Code *object, uint16_t index, uint64_t t)
 {
     object->code(index) = Atom::Timestamp();
     object->code(++index)=Atom(t>>32);
